rename data_synthesize folder to evaluation
