+++ scripts/evaluation/write_lunary_runs.py
@@ -1,6 +1,6 @@
 import sys
 sys.path.append("./")
-from src.dataset_synthesize import lunary_handler
+from src.evaluation import lunary_handler
 from src.mongodb_handler import MongoDBHandler
 import os
 from scripts.scripts_utils import get_config
